refactor(033): move pig IPC side effects from App into usePigMovement
App was still calling setPigDragActive and subscribing to gather-pigs /
display-region directly. CLAUDE.md requires components to be view-only
and hooks to own state + effects. Hook now exposes setDragActive and
manages both subscriptions internally.
